* (OSX) Fixed RTBareBones example, the xcode project needed ArcadeInputComponent.cpp added * Stripped the windows specific stuff from the Irrlicht code, so Windows GL can be used for any platform (we didn't want it, Proton handles all the platform specifics) * (OSX) RT3DApp project now compiles for OSX (GL) * (OSX) Tweaked the shared objc code to stop a crash on exit that was happening, only seemed to happen on the RT3DApp example though, hrm. * (Fmod audio target) Small tweaks, now only supports the latest FMOD version

+++ proton/shared/Entity/DPadComponent.cpp
@@ -17,8 +17,8 @@ void DPADButton::OnButtonChange(int key, bool bDown)
 		
 		if (!m_bDown)
 		{
-			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR, key , bDown);  
-			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW, key , bDown);  
+			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR, (float)key , (float)bDown);  
+			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW, (float)key , (float)bDown);  
 //			LogMsg("DOWN: Got %d down is %d", key, int(bDown));
 			m_bDown = true;
 		}
@@ -28,7 +28,7 @@ void DPADButton::OnButtonChange(int key, bool bDown)
 		if (m_bDown)
 		{
 			m_bDown = false;
-			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW, key , bDown);  
+			GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW, (float)key , (float)bDown);  
 //			LogMsg("RELEASE: Got %d down is %d", key, int(bDown));
 		}
 	}
